Update README.md
The starting point for this project is the agony of choice that every person has almost every day. In many of our everyday situations, we have to make decisions that are not so easy for us. Some of these decisions are made consciously, others unconsciously. However, the decisions we make consciously are based on certain decision criteria. So there is a certain process going on in our heads. I want to build my project on this process in order to organize and structure the thoughts we have and to weigh up not only positive but also negative points from each other. I want to create an instrument that will contribute to individual decision-making. Whether the decision is right or wrong is not to be made in the context of this work. It is only intended to provide assistance to make decisions more easily on the basis of self-imposed points.